update faculty id from into to long

+++ users-microservice/src/main/java/nl/tudelft/sem/template/users/domain/RegistrationService.java
@@ -76,7 +76,7 @@ public class RegistrationService {
      * @param assignedFaculty the assigned faculty of the user.
      * @return the newly created faculty account.
      */
-    public FacultyAccount addFacultyAccount(String netId, int assignedFaculty) {
+    public FacultyAccount addFacultyAccount(String netId, long assignedFaculty) {
         FacultyAccount newFacultyAccount = new FacultyAccount(netId, assignedFaculty);
         facultyAccountRepository.save(newFacultyAccount);
         return newFacultyAccount;
